Option to keep logged in

+++ src/shared/services/authentication.service.ts
@@ -14,7 +14,14 @@ export class AuthenticationService {
   public currentUser: Observable<User>;
 
   constructor(private apiService: ApiService) {
-    this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(localStorage.getItem('currentUser')));
+    if (localStorage.getItem('currentUser') && JSON.parse(localStorage.getItem('currentUser')) !== {}) {
+      console.log('current user from local')
+      this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(localStorage.getItem('currentUser')));
+    } else {
+      console.log('current user from session')
+      this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(sessionStorage.getItem('currentUser')));
+    }
+    console.log('curren user: ' + this.currentUserSubject.value)
     this.currentUser = this.currentUserSubject.asObservable();
   }
 
@@ -22,14 +29,21 @@ export class AuthenticationService {
     return this.currentUserSubject.value;
   }
 
-  login(username: string, password: string) {
+  login(username: string, password: string, keepLoggedIn: Boolean) {
+    console.log("KeepLoggedIn: " + keepLoggedIn)
     return this.apiService.post(`/user/signin`, { username, password })
       .map(data => {
         const user = data.user;
         // login successful if there's a jwt token in the response
         if (user && user.token) {
-          // store user details and jwt token in local storage to keep user logged in between page refreshes
-          localStorage.setItem('currentUser', JSON.stringify(user));
+          if (keepLoggedIn) {
+            console.log('keep logged in')
+            // store user details and jwt token in local storage to keep user logged in between page refreshes
+            localStorage.setItem('currentUser', JSON.stringify(user));
+          } else {
+            console.log('not keep logged in')
+            sessionStorage.setItem('currentUser', JSON.stringify(user));
+          }
           this.currentUserSubject.next(user);
         }
